+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+- Support comma+space-separated headers in check-http.rb
+- Support POST requests in check-http.rb
+- Add a Test Kitchen config and BATS tests for CI
 
 ## [0.2.1] - 2015-12-14
 ### Added